[Taxon][Behat] Extract management of image to separate form element.
