+++ src/Sylius/Behat/Context/Ui/Admin/ManagingTaxonsContext.php
@@ -17,6 +17,7 @@ use Behat\Behat\Context\Context;
 use FriendsOfBehat\PageObjectExtension\Page\SymfonyPageInterface;
 use Sylius\Behat\Context\Ui\Admin\Helper\ValidationTrait;
 use Sylius\Behat\Element\Admin\Taxon\FormElementInterface;
+use Sylius\Behat\Element\Admin\Taxon\FormImageElementInterface;
 use Sylius\Behat\NotificationType;
 use Sylius\Behat\Page\Admin\Product\UpdateSimpleProductPageInterface;
 use Sylius\Behat\Page\Admin\Taxon\CreateForParentPageInterface;
@@ -40,6 +41,7 @@ final class ManagingTaxonsContext implements Context
         private readonly CreateForParentPageInterface $createForParentPage,
         private readonly UpdatePageInterface $updatePage,
         private readonly FormElementInterface $formElement,
+        private readonly FormImageElementInterface $formImageElement,
         private readonly CurrentPageResolverInterface $currentPageResolver,
         private readonly NotificationCheckerInterface $notificationChecker,
         private readonly JavaScriptTestHelper $testHelper,
@@ -307,7 +309,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function iAttachImageWithType(string $path, ?string $type = null): void
     {
-        $this->formElement->attachImage($path, $type);
+        $this->formImageElement->attachImage($path, $type);
     }
 
     /**
@@ -323,7 +325,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function thisTaxonShouldHaveAnImageWithType(string $type): void
     {
-        Assert::true($this->formElement->isImageWithTypeDisplayed($type));
+        Assert::true($this->formImageElement->isImageWithTypeDisplayed($type));
     }
 
     /**
@@ -331,7 +333,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function thisTaxonShouldNotHaveAnImageWithType($code)
     {
-        Assert::false($this->formElement->isImageWithTypeDisplayed($code));
+        Assert::false($this->formImageElement->isImageWithTypeDisplayed($code));
     }
 
     /**
@@ -339,7 +341,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function iRemoveAnImageWithType(string $type): void
     {
-        $this->formElement->removeImageWithType($type);
+        $this->formImageElement->removeImageWithType($type);
     }
 
     /**
@@ -347,7 +349,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function iRemoveTheFirstImage(): void
     {
-        $this->formElement->removeFirstImage();
+        $this->formImageElement->removeFirstImage();
     }
 
     /**
@@ -357,7 +359,7 @@ final class ManagingTaxonsContext implements Context
     {
         $this->iWantToModifyATaxon($taxon);
 
-        Assert::same($this->formElement->countImages(), 0);
+        Assert::same($this->formImageElement->countImages(), 0);
     }
 
     /**
@@ -365,7 +367,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function iChangeItsImageToPathForTheType($path, $type): void
     {
-        $this->formElement->changeImageWithType($type, $path);
+        $this->formImageElement->changeImageWithType($type, $path);
     }
 
     /**
@@ -373,7 +375,7 @@ final class ManagingTaxonsContext implements Context
      */
     public function iChangeTheFirstImageTypeTo($type): void
     {
-        $this->formElement->modifyFirstImageType($type);
+        $this->formImageElement->modifyFirstImageType($type);
     }
 
     /**
@@ -384,7 +386,7 @@ final class ManagingTaxonsContext implements Context
     {
         $this->iWantToModifyATaxon($taxon);
 
-        Assert::same($this->formElement->countImages(), (int) $count);
+        Assert::same($this->formImageElement->countImages(), (int) $count);
     }
 
     /**
